requirements.txt
Flask==0.12.2
Flask-WTF==0.14.2
Jinja2==2.9.4
MarkupSafe==0.23
WTForms==2.1
Werkzeug==0.11.15
click==6.7
gunicorn==19.6.0
itsdangerous==0.24
wsgiref==0.1.2
Flask-Migrate
Flask_script
Flask_uploads
Flask_images
psycopg2
Flask_MySql
Flask_mysqldb
Flask_Login
Faker == 1.8.7